Add dithering to resampler
We use a simple LCG to generate a uniform(0, ULP) variable which is
added to accumulators before they are truncated.  This decreases
distortion (quantization noise) at the cost of a higher noise floor.
The LCG is chosen because it is easy to vectorize.

Dithering has almost no impact on performance.