Rework universal binary build system
The new system uses a wrapper script which calls the main configure
script several times.  The makefile is split into a part which gets
duplicated for each architecture (arch.mak.in) and another part which
doesn't (Makefile.in).  Make is not used recursively, so this new system
should not run into those particular edge cases.